agent-swarm/requirements.txt
openai>=1.3.0
pydantic>=2.0.0
rich>=12.0.0
python-dotenv>=0.19.0
asyncio
aiofiles>=23.0.0
pyyaml>=6.0.0
uuid
datetime
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0